use new prettyprinting method for protobuf
