+++ src/main/scala/protobuf/schema.scala
@@ -18,7 +18,6 @@ package skeuomorph
 package protobuf
 
 import cats.Functor
-import qq.droste.Algebra
 
 sealed trait ProtobufF[A]
 object ProtobufF {
@@ -80,62 +79,6 @@ object ProtobufF {
           fields.map(field => field.copy(tpe = f(field.tpe))),
           reserved
         )
-
     }
-  }
-
-  def printOption(o: Option): String = s"${o.name} = ${o.value}"
-
-  def render: Algebra[ProtobufF, String] = Algebra {
-    case TDouble()        => "double"
-    case TFloat()         => "float"
-    case TInt32()         => "int32"
-    case TInt64()         => "int64"
-    case TUint32()        => "uint32"
-    case TUint64()        => "uint64"
-    case TSint32()        => "sint32"
-    case TSint64()        => "sint64"
-    case TFixed32()       => "fixed32"
-    case TFixed64()       => "fixed64"
-    case TSfixed32()      => "sfixed32"
-    case TSfixed64()      => "sfixed64"
-    case TBool()          => "bool"
-    case TString()        => "string"
-    case TBytes()         => "bytes"
-    case TNamedType(name) => name
-
-    case TRequired(value) => s"required $value"
-    case TOptional(value) => s"optional $value"
-    case TRepeated(value) => s"repeated $value"
-
-    case TEnum(name, symbols, options, aliases) =>
-      val printOptions = options.map(o => s"option ${o.name} = ${o.value}").mkString("\n")
-      val printSymbols = symbols.map({ case (s, i) => s"$s = $i;" }).mkString("\n")
-      val printAliases = aliases.map({ case (s, i) => s"$s = $i;" }).mkString("\n")
-      s"""
-      |enum $name {
-      |  $printOptions
-      |  $printSymbols
-      |  $printAliases
-      |}
-      """.stripMargin
-    case TMessage(name, fields, reserved) =>
-      val printReserved = reserved.map(l => s"reserved " + l.mkString(", ")).mkString("\n  ")
-      def printOptions(options: List[Option]) =
-        if (options.isEmpty)
-          ""
-        else
-          options.map(printOption).mkString(start = " [", sep = ", ", end = "]")
-
-      val printFields =
-        fields
-          .map(f => s"${f.tpe} ${f.name} = ${f.position}${printOptions(f.options)};")
-          .mkString("\n  ")
-      s"""
-      |message $name {
-      |  $printReserved
-      |  $printFields
-      |}
-      """.stripMargin
   }
 }
